Remove references to xdoc_1_0.dtd.
Some of the site xdoc documents had a DOCTYPE declaration pointing to
the DTD at

   PUBLIC "-//Apache Software Foundation//DTD XDOC 1.0//EN"
          "http://maven.apache.org/dtd/xdoc_1_0.dtd"

This URL seems to have gone 404.  Just remove the declaration.
It does mean that I no longer know what the document format is though.
Quick google attempts to grub up a copy of the DTD failed.