change version number format to be more maven-friendly
I've replaced version 1.3-4+ with 1.3.4-post.

So the third number now comes after a "-" not a ".".

I'm using "-post" instead of "+" as a suffix to indicate something
after the given numbered version.  This string is not magic to maven,
but it is interpreted as a build number.  I don't plan to update it
after every change unless it becomes clear that there's some good
reason to do that.  If I'm leading up to a release I will use
"-SNAPSHOT" rather than what I was using before, "-".

This change made following Paul Harrison's comments, with reference to
http://docs.oracle.com/middleware/1212/core/MAVEN/maven_version.htm