Fix web profile GUI so it doesn't get too wide
The part of the Web Profile popup dialogue that displays client
Name, Origin and URL is now limited in width, so the popup window
doesn't get ridiculously wide.  You can still scroll to see the
whole of long entries by clicking on the text fields and hitting
the right arrow, though there is no visual feedback for this so
it's not obvious.