fix hub to make most client calls without waiting for HTTP response
Change the XmlRpcCallableClient implementation used by the hub to use
callAndForget rather than callAndWait.  The effect of this is that
hub call() and callAll() invocations do not wait for the completion
of the XML-RPC call (the HTTP response) before returning from the
client's call.

This sounds like sense, and it's the way it used to happen in JSAMP 1.1
(before Feb 14, 2011, commit 764a76e39).  I made the change from
callAndForget to callAndWait that time because it was causing trouble
with hub shutdown during JVM shutdown - shutdown messages to the
clients were not getting completed.  However, with the improved
JVM shutdown management that is now in place, I think it the current
change may manage to work OK in both cases.  We'll see.