generalise credential presentation API
The logic of how to extract user-facing credential information
from the HTTP header and explicitly presented securityMap is
extracted from HubSwingClientAuthorizer into an external
class WebCredentialPresenter, implementing new interface
CredentialPresenter.

When presenting credentials, the HTTP request and the securityMap
explicitly provided by the registering application are examined
(previously, it was the request and just the samp.name extracted
from that map).  This makes it possible to look at more general
credentials.

Also, instead of returning a boolean to indicate success, the
credential examination methods now have a void return and either
complete (success) or throw a SampException to indicate failure.
That allows better feedback of the reason for registration failure
(the exception should have a user-readable message), and also
simplifies some of the implementation.

Finally, took the opportunity to update some document references
(the Web Origin document is now available as RFC 6454).