log tlsamp-related changes in history file
Various changes have been merged from the "tls" branch, introduced
as required for implementing the experimental TLS (HTTPS) Profile.
That profile is not included here, it's a separate project.

Most of the changes are harmless, largely publicising existing
methods for external use.  One change to the API is that the
ClientAuthorizer.authorize method now returns void and throws
an exception rather than returning boolean.  This is a
backwardly incompatible change to the public API, but I doubt
that other code is invoking this method, so I don't expect
it to cause client problems.