add jsamp.notray system property
Setting jsamp.notray=true will prevent installation of the hub
into the System Tray even if java.awt.SystemTray reports
system tray support.

In some cases it seems that Gnome claims to support the system
tray but doesn't, which means the hub menus can be inaccessible.
This was at least a historical problem; it's not clear if it
still applies in recent Gnome versions, but it doesn't hurt to
have the configuration option present just in case.