Update downloads page
The FTP site has been replaced by HTTPS, and the Astrogrid maven repo
is now an online gambling page.  Add a link to the maven central repo.
I no longer have a code-signing certificate, so remove reference
to the signed jar file.  Hopefully this is not going to inconvenience
people.