Remove finalizer from XmlRpcHubConnection
Finalizers are deprecated at Java 9 for eventual removal from
the JDK, and are generally problematic.
Remove the finalizer that attempts to unregister the
XmlRpcHubConnection if it has not already been unregistered
once it becomes unreachable.  Client code ought to unregister
explicitly anyway, and if it doesn't, this should get done at
JVM shutdown time anyway; finalization is in any case not
guaranteed by the JVM.  So this is unlikely to make much
difference to behaviour.